feat: fixed routes and config files
