+++ src/pages/portfolio/useful-links/_config/columns/facetedFilters.tsx
@@ -1,0 +1,18 @@
+import { ITableFacetedFilter } from '@/types';
+
+export const type1FacetedFilters: ITableFacetedFilter[] = [
+	{
+		id: 'status',
+		title: 'Status',
+		options: [
+			{
+				label: 'Success',
+				value: 'success',
+			},
+			{
+				label: 'Failed',
+				value: 'failed',
+			},
+		],
+	},
+];
